opendata_module/anonymizer/requirements.txt
psycopg2==2.7.1
pymongo==3.3.0
python-dateutil==2.6.0
pyyaml==3.12
